fix(compat): compile netstandard core
Add compatibility shims and guarded fallbacks so the core library builds under the forced netstandard2.0 probe while keeping net10 paths intact.

Refs #1300

+++ src/Dekaf/Producer/Partitioner.cs
@@ -49,7 +49,7 @@ public sealed class DefaultPartitioner : IPartitioner
 public sealed class StickyPartitioner : IPartitioner, IBatchCompletionAwarePartitioner
 {
     private readonly ConcurrentDictionary<string, int> _stickyPartitions = new();
-    private uint _counter;
+    private int _counter;
 
     public int Partition(string topic, ReadOnlySpan<byte> key, bool keyIsNull, int partitionCount)
     {
@@ -64,7 +64,7 @@ public sealed class StickyPartitioner : IPartitioner, IBatchCompletionAwareParti
             // Cold path: topic not yet seen, compute and add a partition.
             // Use uint to avoid overflow to negative values.
             // GetOrAdd handles the race condition - if another thread added first, we use their value.
-            var newPartition = (int)(Interlocked.Increment(ref _counter) % (uint)partitionCount);
+            var newPartition = (int)((uint)Interlocked.Increment(ref _counter) % (uint)partitionCount);
             return _stickyPartitions.GetOrAdd(topic, newPartition);
         }
 
@@ -80,8 +80,8 @@ public sealed class StickyPartitioner : IPartitioner, IBatchCompletionAwareParti
         // Use uint to avoid overflow to negative values.
         _stickyPartitions.AddOrUpdate(
             topic,
-            _ => (int)(Interlocked.Increment(ref _counter) % (uint)partitionCount),
-            (_, _) => (int)(Interlocked.Increment(ref _counter) % (uint)partitionCount));
+            _ => (int)((uint)Interlocked.Increment(ref _counter) % (uint)partitionCount),
+            (_, _) => (int)((uint)Interlocked.Increment(ref _counter) % (uint)partitionCount));
     }
 }
